# microservices/mano-dns/mano-dns-pdns-aggregator/mano-pdns-api/src/main/java/com/ubiqube/mano/pdns/model/Autoprimary.java
/*
 * PowerDNS Authoritative HTTP API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.15
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

package com.ubiqube.mano.pdns.model;

import java.util.Objects;

import org.jspecify.annotations.Nullable;

import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;

/**
 * An autoprimary server that can provision new domains.
 */
@JsonPropertyOrder({
		Autoprimary.JSON_PROPERTY_IP,
		Autoprimary.JSON_PROPERTY_NAMESERVER,
		Autoprimary.JSON_PROPERTY_ACCOUNT
})
@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2025-05-23T10:34:53.990305454+02:00[Europe/Paris]", comments = "Generator version: 7.13.0")
public class Autoprimary {
	public static final String JSON_PROPERTY_IP = "ip";
	@Nullable
	private String ip;

	public static final String JSON_PROPERTY_NAMESERVER = "nameserver";
	@Nullable
	private String nameserver;

	public static final String JSON_PROPERTY_ACCOUNT = "account";
	@Nullable
	private String account;

	public Autoprimary() {
	}

	public Autoprimary ip(@Nullable final String ip) {

		this.ip = ip;
		return this;
	}

	/**
	 * IP address of the autoprimary server
	 *
	 * @return ip
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_IP)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getIp() {
		return ip;
	}

	@JsonProperty(JSON_PROPERTY_IP)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setIp(@Nullable final String ip) {
		this.ip = ip;
	}

	public Autoprimary nameserver(@Nullable final String nameserver) {

		this.nameserver = nameserver;
		return this;
	}

	/**
	 * DNS name of the autoprimary server
	 *
	 * @return nameserver
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_NAMESERVER)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getNameserver() {
		return nameserver;
	}

	@JsonProperty(JSON_PROPERTY_NAMESERVER)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setNameserver(@Nullable final String nameserver) {
		this.nameserver = nameserver;
	}

	public Autoprimary account(@Nullable final String account) {

		this.account = account;
		return this;
	}

	/**
	 * Account name for the autoprimary server
	 *
	 * @return account
	 */
	@Nullable
	@JsonProperty(JSON_PROPERTY_ACCOUNT)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

	public String getAccount() {
		return account;
	}

	@JsonProperty(JSON_PROPERTY_ACCOUNT)
	@JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
	public void setAccount(@Nullable final String account) {
		this.account = account;
	}

	@Override
	public boolean equals(final Object o) {
		if (this == o) {
			return true;
		}
		if ((o == null) || (getClass() != o.getClass())) {
			return false;
		}
		Autoprimary autoprimary = (Autoprimary) o;
		return Objects.equals(this.ip, autoprimary.ip) &&
				Objects.equals(this.nameserver, autoprimary.nameserver) &&
				Objects.equals(this.account, autoprimary.account);
	}

	@Override
	public int hashCode() {
		return Objects.hash(ip, nameserver, account);
	}

	@Override
	public String toString() {
		StringBuilder sb = new StringBuilder();
		sb.append("class Autoprimary {\n");
		sb.append("    ip: ").append(toIndentedString(ip)).append("\n");
		sb.append("    nameserver: ").append(toIndentedString(nameserver)).append("\n");
		sb.append("    account: ").append(toIndentedString(account)).append("\n");
		sb.append("}");
		return sb.toString();
	}

	/**
	 * Convert the given object to string with each line indented by 4 spaces
	 * (except the first line).
	 */
	private String toIndentedString(final Object o) {
		if (o == null) {
			return "null";
		}
		return o.toString().replace("\n", "\n    ");
	}

}