# microservices/mano-dns/mano-dns-pdns-aggregator/mano-pdns-api/src/main/resources/authoritative-api-swagger.yaml
swagger: '2.0'
info:
  version: "0.0.15"
  title: PowerDNS Authoritative HTTP API
  license:
    name: MIT
basePath: /api/v1
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  # X-API-Key: abcdef12345
  APIKeyHeader:
    type: apiKey
    in: header
    name: X-API-Key
security:
  - APIKeyHeader: []

# Overall TODOS:
# TODO: Return types are not consistent across documentation
#       We need to look at the code and figure out the default HTTP response
#       codes and adjust docs accordingly.
paths:
...
  '/servers/{server_id}/autoprimaries':
    parameters:
      - name: server_id
        in: path
        required: true
        description: 'The id of the server to manage the list of autoprimaries on'
        type: string
    get:
      summary: 'Get a list of autoprimaries'
      operationId: getAutoprimaries
      tags:
        - autoprimary
      responses:
        '200':
          description: OK.
          schema:
            $ref: '#/definitions/Autoprimary'
        <<: *commonErrors
    post:
      summary: 'Add an autoprimary'
      description: 'This methods add a new autoprimary server.'
      operationId: createAutoprimary
      tags:
        - autoprimary
      parameters:
        - name: autoprimary
          description: autoprimary entry to add
          required: true
          in: body
          schema:
            $ref: '#/definitions/Autoprimary'
      responses:
        '201':
          description: Created
        <<: *commonErrors
  '/servers/{server_id}/autoprimaries/{ip}/{nameserver}':
    parameters:
      - name: server_id
        in: path
        required: true
        description: 'The id of the server to delete the autoprimary from'
        type: string
      - name: ip
        in: path
        required: true
        description: 'IP address of autoprimary'
        type: string
      - name: nameserver
        in: path
        required: true
        description: 'DNS name of the autoprimary'
        type: string
    delete:
      summary: 'Delete the autoprimary entry'
      operationId: deleteAutoprimary
      tags:
        - autoprimary
      responses:
        '204':
          description: 'OK, key was deleted'
        <<: *commonErrors
definitions:
...
  Autoprimary:
     title: Autoprimary server
     description: An autoprimary server that can provision new domains.
     properties:
       ip:
         type: string
         description: "IP address of the autoprimary server"
       nameserver:
         type: string
         description: "DNS name of the autoprimary server"
       account:
         type: string
         description: "Account name for the autoprimary server"
...